+++ src/Macros/Commands/Context/CopyToGlobalCommand.cs
@@ -19,6 +19,7 @@ internal sealed class CopyToGlobalCommand : BaseCommand<CopyToGlobalCommand>
     {
         var current = MacroSelectionContext.Current;
         Command.Visible = current?.Scope == MacroScope.Repo;
+        Command.Supported = Command.Visible;
     }
 
     /// <inheritdoc />